pyproject.toml
[project]
name = "obsidian-podcast"
version = "0.1.0"
description = "RSS feed articles to audio for Obsidian"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiofiles>=24.0",
    "feedparser>=6.0",
    "httpx>=0.27.0",
    "readability-lxml>=0.8",
    "lxml[html_clean]>=5.0",
    "beautifulsoup4>=4.12",
    "langdetect>=1.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "alkana>=0.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
    "respx>=0.22.0",
]

[project.scripts]
obsidian-podcast = "obsidian_podcast.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_podcast"]

[tool.pytest.ini_options]
testpaths = ["src"]
pythonpath = ["src"]

[tool.ruff]
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
